src/test/java/RegisterServiceImplTests.java
import static org.junit.Assert.*;
import static org.mockito.Mockito.*;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.web.DAOInterface.RegisterDaoInterface;
import com.web.Entity.Book;
import com.web.Service.serviceImpl.RegisterServiceImpl;

public class RegisterServiceImplTests {

    @Mock
    private RegisterDaoInterface registerDaoMock;

    @InjectMocks
    private RegisterServiceImpl registerService;

    @BeforeClass
    public static void setUpBeforeClass() throws Exception {
        System.out.println("Before all tests");
    }

    @AfterClass
    public static void tearDownAfterClass() throws Exception {
        System.out.println("After all tests");
    }

    @Before
    public void setUp() throws Exception {
        // Initialize mocks
        MockitoAnnotations.openMocks(this);
        System.out.println("Before each test");
    }

    @After
    public void tearDown() throws Exception {
        System.out.println("After each test");
    }

    @Test
    public void testRegisterServiceImpl() {
        assertNotNull("RegisterServiceImpl object should be created", registerService);
        System.out.println("testRegisterServiceImpl executed");
    }

    @Test
    public void testAddUser() {
        // Arrange
        Object user = new Object(); // Replace with an actual user entity if available
        String userType = "buyer";
        when(registerDaoMock.addUser(user, userType)).thenReturn("User added successfully");

        // Act
        String result = registerService.addUser(user, userType);

        // Assert
        assertEquals("User added successfully", result);
        verify(registerDaoMock, times(1)).addUser(user, userType);
        System.out.println("testAddUser result: " + result);
    }

    @Test
    public void testAuthenticateUser() {
        // Arrange
        String email = "test@example.com";
        String password = "password123";
        Object expectedUser = new Object(); // Replace with an actual user entity if available
        when(registerDaoMock.authenticateUser(email, password)).thenReturn(expectedUser);

        // Act
        Object result = registerService.authenticateUser(email, password);

        // Assert
        assertEquals(expectedUser, result);
        verify(registerDaoMock, times(1)).authenticateUser(email, password);
        System.out.println("testAuthenticateUser result: " + result);
    }

    @Test
    public void testRegisterBook() {
        // Arrange
        Book book = new Book(); // Ensure this Book object has the required fields filled
        Long retailerId = 1L;
        when(registerDaoMock.registerBook(book, retailerId)).thenReturn("Book registered successfully");

        // Act
        String result = registerService.registerBook(book, retailerId);

        // Assert
        assertEquals("Book registered successfully", result);
        verify(registerDaoMock, times(1)).registerBook(book, retailerId);
        System.out.println("testRegisterBook result: " + result);
    }
}
